refactor: move some behavior from 'publish' task to 'common'
I'm currently working on pubtools-pulp-push. The push task will need to
publish repos and flush caches and ideally would use exactly the same
arguments and behavior as the existing publish task.

Let's refactor a little so that most of the publish logic can be easily
shared across multiple tasks.
